requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
mssql-django==1.7
pyodbc==5.3.0
